refactor: Move ValueBuilder into HaikuComponent
