+++ packages/haiku-common/src/math/geometryUtils.ts
@@ -3,7 +3,7 @@
  */
 
 import HaikuElement from '@haiku/core/lib/HaikuElement';
-import {cubicBezierSplit, distance, Vec2} from '@haiku/core/lib/helpers/PathUtil';
+import {distance, Vec2} from '@haiku/core/lib/helpers/PathUtil';
 import SVGPoints from '@haiku/core/lib/helpers/SVGPoints';
 import Layout3D from '@haiku/core/lib/Layout3D';
 import create from '@haiku/core/lib/vendor/gl-mat4/create';
