+++ packages/haiku-serialization/src/bll/TimelineProperty.js
@@ -1,6 +1,5 @@
 const {getFallback} = require('@haiku/core/lib/HaikuComponent')
 const logger = require('haiku-serialization/src/utils/LoggerInstance')
-require('@haiku/core/lib/ValueBuilder') // Implicit dependency; do not remove
 
 const TimelineProperty = {}
 
@@ -169,17 +168,14 @@ TimelineProperty.getPropertyValueAtTime = (
 
   if (propertiesGroup) {
     try {
-      // The hostInstance, which should be a HaikuCore, should have a 'ValueBuilder' attached to it under the property name 'builder'
-      // This instance is responsible for dependency injection, caching, and recalc of transitioning values on the fly. (Pardon the dumb name.)
-      if (hostInstance && hostInstance.builder) {
-        const computedValue = hostInstance.builder.grabValue(
+      if (hostInstance) {
+        const computedValue = hostInstance.grabValue(
           timelineName,
           componentId,
           hostInstance.findElementsByHaikuId(componentId)[0],
           outputName, // propertyName
           propertiesGroup[outputName], // propertyValue
           time,
-          hostInstance, // haikuComponent
           !hostInstance.shouldPerformFullFlush(), // isPatchOperation
           true, // skipCache
           false // clearSortedKeyframesCache
